refactor(core): split 3 god files into SRP-focused packages (#180)
Split dependencies.py (427), event_bus.py (542), and llm.py (638) into
focused modules following Single Responsibility Principle.

app/dependencies.py → app/dependencies/ (5 modules + hub)
  - auth.py: get_current_user, oauth2_scheme, role checks
  - cross_tenant.py: cross-tenant pre-checks, admin factories
  - db_deps.py: get_db_with_tenant, DB session deps
  - event_deps.py: EventBus singleton factory
  - llm_deps.py: LLM provider dependency

app/event_bus.py → app/event_bus/ (3 modules + hub)
  - bus.py: EventBus publisher class
  - consumer.py: EventConsumer consumer group class
  - _helpers.py: DLQ tracking, retry keys, drain

app/core/llm.py → app/core/llm/ (3 modules + hub)
  - config.py: retry config, credential redaction
  - providers.py: LLMProvider base + OpenAI/Anthropic/Gemini
  - factory.py: provider factory, singletons, registry

Hub __init__.py files re-export all symbols for backward compatibility.
488 unit tests pass, 0 failures.
