+++ app/event_bus/__init__.py
@@ -1,0 +1,20 @@
+"""Event bus abstraction for Redis Streams.
+
+Provides EventBus (publisher) and EventConsumer (consumer group) classes
+that wrap Redis Streams primitives with typed Pydantic event schemas.
+"""
+from __future__ import annotations
+
+from app.event_bus._helpers import (  # noqa: F401
+    _INFLIGHT_DLQ,
+    _RETRY_COUNT_KEY,
+    _RETRY_HASH_FIELD,
+    _retry_key,
+    drain_dlq_tasks,
+)
+from app.event_bus.bus import (  # noqa: F401
+    EventBus,
+)
+from app.event_bus.consumer import (  # noqa: F401
+    EventConsumer,
+)
